ПР14

+++ controlers/cards.js
@@ -1,5 +1,10 @@
 const Card = require("../models/card");
 
+const BadRequestError = require("../middleware/BadRequestError");
+const NotFoundError = require("../middleware/NotFoundError");
+
+const ForbiddenError = require("../middleware/ForbiddenError");
+
 const getCard = async (req, res) => {
     try {
         const cards = await Card.find({});
@@ -12,14 +17,19 @@ const getCard = async (req, res) => {
 const deleteCardById = async (req, res) => {
     try {
         const card = await Card.findById(req.params.cardId);
+
         if (!card) {
-            res.status(404).send({ message: "Карточка не найдена" });
+            throw new NotFoundError("Карточка не найдена");
+        }
+
+        if (card.owner !== req.user._id) {
+            throw new ForbiddenError("Невозможно удалить карточку");
         }
         await Card.findByIdAndRemove(req.params.cardId);
         res.send({ message: "Карточка удалена" });
     } catch (err) {
         if (err.name === "CastError") {
-            res.status(400).send({ message: "Невалидные данные" });
+            throw new BadRequestError("Невалидные данные");
         }
         res.status(500).send({ message: "Ошибка на сервере" });
     }
@@ -29,17 +39,16 @@ const createCard = async (req, res) => {
     try {
         const { name, link } = req.body;
         if (!name || !link) {
-            res.status(400).send({ message: "Переданны неверные данные" });
-            return;
+            throw new BadRequestError("Переданны неверные данные");
         }
         const card = await Card.create({ name, link, owner: req.user._id });
         if (!card) {
-            res.status(404).send({ message: "Карточка не создана" });
+            throw new NotFoundError("Карточка не создана");
         }
         res.status(201).send(card);
     } catch (err) {
         if (err.name === "ValidationError") {
-            res.status(400).send({ message: "Невалидные данные" });
+            throw new BadRequestError("Невалидные данные");
         }
         res.status(500).send({ message: "Ошибка на сервере" });
     }
@@ -53,12 +62,12 @@ const likeCard = async (req, res) => {
             { new: true },
         );
         if (!card) {
-            res.status(404).send({ message: "Карточка не найдена" });
+            throw new NotFoundError("Карточка не найдена");
         }
         res.send(card);
     } catch (err) {
         if (err.name === "CastError") {
-            res.status(400).send({ message: "Невалидные данные" });
+            throw new BadRequestError("Невалидные данные");
         }
         res.status(500).send({ message: "Ошибка на сервере" });
     }
@@ -72,12 +81,12 @@ const dislikeCard = async (req, res) => {
             { new: true },
         );
         if (!card) {
-            res.status(404).send({ message: "Карточка не найдена" });
+            throw new NotFoundError("Карточка не найдена");
         }
         res.send(card);
     } catch (err) {
         if (err.name === "CastError") {
-            res.status(400).send({ message: "Невалидные данные" });
+            throw new BadRequestError("Невалидные данные");
         }
         res.status(500).send({ message: "Ошибка на сервере" });
     }
